backend/requirements.txt
# Django & DRF
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.11.0
django-redis>=5.3.0
django-ratelimit>=4.1.0

# Database
psycopg2-binary>=2.9.0

# Kotak Neo API
neo-api-client>=1.1.0

# WebSocket & Async
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
websockets>=11.0.0

# Celery
celery>=5.3.0
redis>=4.6.0

# Security & Encryption
cryptography>=41.0.0
pyotp>=2.9.0

# Utilities
python-decouple>=3.8
requests>=2.31.0
python-dateutil>=2.8.0
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0
